# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tchallenger",
        "description": "",
        "contact": {
            "email": "nathalietaffot@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/public/api/parametres": {
            "get": {
                "tags": [
                    "Parameter"
                ],
                "summary": "Get parameter",
                "description": "Get parameter",
                "operationId": "indexxxx",
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "affichage des parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            []
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parameter"
                ],
                "summary": "create Parameter",
                "description": "create Parameter",
                "operationId": "storeeeee",
                "requestBody": {
                    "description": "Enregistrement d'un nouvel Parameter",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "color": {
                                        "description": "votre login",
                                        "type": "string",
                                        "format": "string",
                                        "example": "laporte"
                                    },
                                    "notif_abonnement": {
                                        "description": "notification abonnement",
                                        "type": "boolean",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "notif_challenge": {
                                        "description": "notification challenge",
                                        "type": "boolean",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "notif_comment": {
                                        "description": "notification comment",
                                        "type": "boolean",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "notif_publication": {
                                        "description": "notification publication",
                                        "type": "boolean",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "notif_message": {
                                        "description": "notification message",
                                        "type": "boolean",
                                        "format": "1",
                                        "example": "1"
                                    },
                                    "langue": {
                                        "description": "votre langue",
                                        "type": "string",
                                        "format": "string",
                                        "example": "fr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Parametre bien Creer."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "400"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du traitement de la demande"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de connexion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            []
                        ]
                    }
                ]
            }
        },
...
    }
}